backend/requirements.txt
flask
flask-cors
torch
torchvision
matplotlib
pillow
numpy
